maria/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
click==8.1.7
distro==1.9.0
docstring_parser==0.16
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
instructor==1.4.0
jiter==0.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==2.1.0
openai==1.42.0
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rich==13.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
yarl==1.9.6
